Final portion of fix for bug 3178
This commit implements the fix for LongAsync payloads.
More testing will be done before this is merged to HEAD.

As with Long (and the Extended API before) the solution is to have a loop
to issue additional calls to gasnetc_rdma_put_bulk() if the first (which
is normally normally the ONLY) returns a short count.  Unlike the case
of the Long we don't block for the RDMA to complete and therefore must
arrange to send the AM header after ALL of the RMDAs have completed even
though that is most likely to happen after we return.  The solution is
to place a weakatomic counter in the (previously otherwise unused) union
portion of the GPD which corresponds to the header.  The logic for GPD
GPD completion is then modified to dec-and-test the counter to get
unambigous identification of the LAST completion.  Note that the counter
is "biased" by 1, and that extra count is decremented after the loop
has issued all the required gasnetc_rdma_put_bulk() calls, and in the
event that all of them had completed, the header is sent at the end of
the loop instead of by the last completion.  This extra compication is
needed because we cannot know until AFTER the last RDMA has been issued
that it actually *is* the last.