Fix bug 3178 for Gemini
I cannot be of certain the problem is specific to the Gemini implementation
of the GNI library, or just to its revision.  Either way, the fix that was
100% effective on Edison did NOT resolve the problem on Grace.  The difference
is that on Grace when the GNI_MemRegister() call failed, it wrote 0s to its
output argument, while on Edison the previous value was preserved.

This commit makes no assuptions about the preservation of the memory handle
output of GNI_MemRegister and unconditionally sets it to the correct value
after a failed call.