Hopefully last pass at cleanups for macro-based multi-domain support
This commit makes changes to make GASNETC_DIDX* macros more similar to
GASNETE_THREAD*, for instance usign a "POST" and an implicit variable
name.

The diffstat shows a near parity of lines added and removed, but the
actual changes remove numerous instances of an explict "didx", which
is now an implicit part of the relevent GASNETC_DIDX_* macros.

As before, this is destined to HEAD but being performed on BUG3178_branch
to limit my testing burden to just the one branch.  If the work to fix
bug 3179 doesn't "make the cut" for the release, then there will be no
harm done by also delaying integration of these cleanups.