Replace bounds checks in critical path with assertions
The code for unaligned gets should be marginally faster by not checking
the validity of the length in an OPT build.

This is another change destinied for HEAD which is be integrated on
the BUG3178_branch to keep my testing to just one branch.  As with
the others, loss of this change is harmless if this branch does not
get merge to HEAD prior to release.