Fix for PSHM support on BG/Q for dirver V1R2M1 or newer.
Prior to the V1R2M1 driver, shm_unlink() was a no-op.
We would work around that by NOT using O_CREAT|O_EXCL on our shm_open()
calls to create our own segments, because those are created and destroyed
as we try to get the size(s) right.  Now that shm_unlink() is working in
the most recent driver, that work-around leads to ENOENT failures.

This commit resolves this problem by retrying with O_CREAT|O_EXCL, but
ONLY on BG/Q and ONLY on a ENOENT failure on creation of ones own segment.
Lacking any preprocesor magic to query the driver version, this is the
probably cleanest solution currently available.