Fix for "starvation" problem in testbarrierconf with "-t"
Some testers (especially those which over-commit cores) have been needlessly
timing out on the "-t" (gasnet_barrier_[t]ry variant) simply because a given
process would spin poll for the entire scheduling quantum before the next one
was given a chance.  This lead to VERY SLOW progress, but not actual deadlock.

This commit adds a yield in the spin-polling loop.