Update PSHM docs with current state of affairs on most recent BG/Q driver.
It looks most likely that this is an IBM bug, but GASNe has not yet been
proven completely blameless.  Even if the fault does lie with IBM, there
is a possibility that GASNet could work around it.  Either way, there is
almost zero chance I will have resources to devote to this between now
and the release.  So, I am documenting the problem.