First commit for branch developing new PSHM implementation for BG/Q
This commit implements PSHM over "GHEAP" (Global HEAP), which is any system
on which virtual addresses from one process are valid in others on the same
compute node.  On BG/Q this is true when BG_MAPCOMMONHEAP=1 in the
environment.  However, this implenetation doesn't include anything which
is specific to BG/Q.