Second commit for branch developing new PSHM implementation for BG/Q
This commit removes the work-arounds for BG/Q-specific "quirks" in the
implementation of POSIX shared memory.  There is no need to maintain
these once this branch is merged to HEAD because we'll no longer be using
POSIX shared memory on BG/Q.  Additionally, POSIX shared memory is not
working correctly as of V1R2M1 (looks like IBM bugs to me).