Add AMRef Get and Put to ibv-conduit
This commit adds configuration defines in ibv-conduit to cause the
AMRef implementations of Get and Put to be built.  This is for some
experiments Eric Roman and I want to perform while working toward the
eventual goal of implementing a Quiesce protcol.

The newly-avalable calls are:
 * gasnete_amref_get_nb_bulk
 * gasnete_amref_get_nbi_bulk
 * gasnete_amref_put_nb_bulk
 * gasnete_amref_put_nbi_bulk
 * gasnete_amref_put_nb
 * gasnete_amref_put_nbi

In a SEQ build (the subject of our intended experiments) all 6 have
the same arguments and semantics as the corresponding gasnet_*() entry
points.  One can build all of the remaining contiguous point-to-point
Get and Put calls from these six plus the following "rules":

1) Form non-bulk Get calls using the corresponding bulk one, since
   the semantics are compatible.

2) Form blocking calls by composing gasnet_wait_syncnb() and the
   proper _nb or _nb_bulk call.

3) Form value-based Puts by passing the following as the 'src' to the
   corresponding address-based Put function:
      GASNETE_STARTOFBITS(&value,nbytes)

4) Form value-based Gets using a temporary of type
   gasnet_register_value_t, initialized to zero and use as 'dest' the
      GASNETE_STARTOFBITS(&temporary,nbytes)
   construct that is for 'src' for value-based Puts.

Examples of all four of these construction rules can be found
somewhere within extended-ref/gasnet_extended.h.

It is worth noting that UPCR does not currently contain non-blocking
value-based gets, which is good since they are annoying to implement
based on the rules above (would need heap-allocated temporaries).