Disable RDMA-Based Longs w/i AMRef for Quiesce work
This is a precautionary measure until we can reason out whether
this is necessary to get the desired ordering propert(y/ies).
However, if this *is* necessary then we probably also need to
do something with respect to any Long AMs in UPCR.  Fortunately
the only ones there now are for the Semaphore code, which is a
Berkeley extension to UPC and therefore out-of-scope for the
current work.