First pass converting tests to new AM interfaces
Notes:

All the AM initiation calls pass "myteam" for their first argument.
That is just fine for now, since the macro expansion of all of AM
calls simply drop it entirely.

Tests of LongAsync have been removed from:
      testam, testcore2, testcore3, testcore4
and we should be adding tests of both Medium and Long with use of
(lc_opt != GASNETEX_LC_INIT) when that support is ready.

The LongAsync call in testnbr has been replaced TEMPORARILY with a
Long with lc_opt = GASNETEX_LC_INIT.  This is semantically a correct
conversion, but we've lost the Async nature as a result.  We should
change the lc_opt argument to use of a real handle when we are able to
implement it correctly.

testcore5 has NOT been updated in this commit.  It is identical to
testcore4 except that it uses the varargs API which we purposefully
exposed, but did not document (as an extension like "test_syncnb" and
similar).  The next commit should add the newer varargs based on
arg-counting macros (C99 required) and convert testcore5.