Parenthesize arguments in gasnet_ammacros.h
As Dan observed, we need to parenthesize the expanded arguments in
all of the AMRequest and AMReply macros.  Without these parens, an
argument that is not a single token could change meaning.  For
instance, for 'c' having type 'char' the expression 'c << 7' will
have only two possible values '0x80' or '0x00'.  However, the AM
macros prior to this commit would produce
   (gasnet_handlerarg_t)c << 7
which has 256 distinct possible values.  The new expansions is
   (gasnet_handlerarg_t)(c << 7)
which gives the desired result.

This commit is semantically a "cherry-pick" of develop branch commit
1d6181eae714ada9a55f963f358eafe8c509b488, but is textually totally
different due to the divergence.