Rename gasnet2ex.h to gasnet.h
Now the one package exposes both "gasnetex.h" and "gasnet.h".  So long
as we are able, the intent is that the later implement as much of the
original "GASNet 1" API as possible in terms of the GASNet-EX API.

Clients which are in the midst of a transition from the older API to
the newer should be able to include both, or just "gasnet.h".  Using
portions of both APIs should be posible.  Some mechanisms still TBD
may be offered to suppress portions of the mapping from old to new,
for the purpose of testing the completeness of the transition.