Stub the LC try and wait operations
This commit adds the lc_handle and lc_group operations to the
extended-ref and to smp- and ibv-conduits.

This implementation is complete for smp-conduit, which performs all
Puts and Long operations synchronously and copies all Medium payloads.

The implementation is "complete" for ibv-conduit and conduits using
the extended-ref (udp and mpi) only because none of them are *yet*
generating any lc_handle other than GASNETEX_INVALID_LC_HANDLE, nor
are they implementing GASNETEX_LC_GROUP.  In otherwords, there is not
currently any way to expose the incompleteness of the lc operations.