New eop/iop formats (ref,smp,ibv)
In both the eop and iop, this commit adds a "flags2" field and expands
the default "threadidx" field to 16-bits.  It also removes the
space-efficient "eopaddr_t" from the eop, and instead forms the free
list by overwriting the "flags", "flags2" and "threadidx" fields with
an actual pointer.

The net result in terms of space is that for an ILP32 platform the
size remains 4-bytes, while for an LP64 platform the size has doubled
to 8-bytes.  This might be just a transient condition, as we are
discussing enlarging the eop to fill a cache line to avoid
false-sharing.