Change name and location of IOPSTATE (ref,ibv)
This commit makes no changes to the local completion *logic* for the
iop in the ref-extended or ibv-conduit.  However, it moves the storage
from the "flags" field to the "flags2" field, and renames all related
identifiers from "IOPSTATE" to "LCSTATE" (with a some additional
changes in names).

These changes are all intended to allow the (as-yet-unwritten) local
completion logic for the eop to share as much as practical with that
of the iop.