Add THREADINFO_OPT for AM Request (smp,ibv,udp,mpi)
This commit makes changes so that the AM Request funtions have the same
"hidden" threadinfo argument as the Extended API functions.

This commit makes these changes in the conduit-independent header
gasnet_ammacros.h, and in smp-, ibv-, udp- and mpi-conduits only.

Some design notes:

+ The variable argument count of AM Request counts mean that the
threadinfo argument can't be last (at least without some serious
pain).  Nor can it be just before the variable arguments unless one
conditionalizes the argument to the va_start() call.  So, it has been
placed between the flags and nargs arguments.

+ No changes are made (nor planned) for Replies since it makes far
more sense to stash the threadinfo in the token than to require
handlers to add a GASNET_BEGIN_FUNCTION() to each handler.  Except for
certain debug checks, the only operation we currently expect to use
thread-specific data in handler context is a Reply.