configure logic to disable non-EX conduits
By default we still probe for all conduits, and if we find support
for a conduit that's not yet been updated to EX we issue an error
suggesting the user download the GASNet-1 distribution.

--disable-<conduit> disables the probe and conduit, bypassing the error
with the conduit disabled.

--enable-<conduit> forces the conduit to be enabled and changes the
fatal error into a warning (leaving a build directory that probably
won't compile). This is mainly for the use of external collaborators
working to implement EX in their conduit.