Move GASNETE_THREAD_FARG and friends to gasnet_help.h
They are also renamed to a GASNETI_* prefix, since they are now
independent of core/extended.

For now gasnet_extended_help.h #defines name-shift macros to also
provide the old GASNETE_ prefix versions, although when convenient we
should rename the rest of the invocations in the extended and remove the
redundant version (I didn't want to create thousands of merge conflicts
for Paul).

Now that eops/iops (and gasnete_threadinfo_t) are changing be fully
shared by the core and extended, we should probably rename all that goop
into the gasneti_ namespace (possibly during the move out of
extended-ref directory).