fixup: Factor handle management out of gasnet_extended.c
Relocate some internal-use-only macros out of a public header and into
an internal one.  Thanks for Dan to pointing out that I'd picked a bad
"parking spot" for these (which belong in a gasnet_extended_internal.h
when that is brought back).  However, if they never do get moved, this
is a far more appropriate place for these to end up.