Move Ops mgmt to new file gasnet_handle_internal.h
This commit creates a new header, gasnet_handle_internal.h which
contains the handle management portions of gasnet_extended_internal.h
and the critical-path operations previously in gasnet_handle.c.

The situation is now such that gasnet_handle.c is no longer included
into gasnet_extended.c.  It is compiled stand-alone (and actaully has
been for awhile).

As a result of this change:
 + The header gasnet_extended_internal.h is once again a candidate for
   per-conduit clones.
 + Some per-conduit definitions intended to configure the handle data
   types and operations have moved from the public/installed headers
   gasnet_{extended,core}_fwd.h of some conduits to the conduit's
   private/non-installed gasnet_core_internal.h header.
   NOTE: The relocated definitions have NOT YET had a corresponding
   change of namespace prefix.