Fix some C++ conformance issues in testcxx
sunCC exposed some conformance issues in testcxx - specifically,
any callback function pointers passed to GASNet (in this case AM handlers
and a backtrace callback) need to be defined using C linkage to ensure
a compatible calling convention when invoked from the C-mode libgasnet.
(In the case of udp-conduit the handler invocation actually occurs in
the C++-mode libamudp, but that code was already correctly enforcing a C
calling convention for AM handlers).

This precision fix decorates the specific problematic callbacks with extern C.
Enforcing an extern C region around the entire test would also have been
sufficient, but would have weakened the point of the test (which is to
demonstrate GASNet use in a "real" C++ client).