Require "opt-in" to LC support by conduits
Any conduit which does not (yet) have LC support should not pay the
cost for the associated generality in the eop and iop.  So, a conduit
must explicitly enable LC for the associated checks to occur in the
test/wait operations.

This commit additionally allows selecting boolean-vs-counted
representation of the EOP's LC event, independent of the root event.