ibv-conduit: callback-based handling of CQ entries
This commit revises the completion event handling in ibv-conduit to
use a callback in place of simply incrementing an atomic counter.
This will allow the completion handling to perform operations which
require access to more of the corresponding eop or iop than just the
completed-op counter (such as the scheduling of dependent operation).

By using callbacks, the *initiator* now encodes which of 11 distinct
(sets of) actions need to be taken without the need for complex
branching in the completion code.

It is also possible that this may lead to elimination of the COUNTED
variant of the EOP since the (initiated==completed) check will soon
reside in the internal-op completion code, rather than the code which
tests an eop.