Update all conduit configure descriptions
The conduit description strings appear in the GASNet configure summary
output, as well as configure --help.

This is an important channel of information to users regarding GASNet's
support for their system, and as such should convey as much useful
information as possible in the limited space, using high-level terms that
should be understandable to end users.

Update all conduit configure descriptions to include more useful information
and follow a unified format:

 ("Portable") (Vendor Name) (Hardware Name) [API Name] "conduit"

where one or more description fields may be omitted, as appropriate.