ibv-conduit is no longer "counted"
All eop completions are now performed in terms of the same event bits
as in the "EOP_BOOLEAN" case, with the counters now entirely a
conduit-specific customization via the few provided hooks.  This
customization consists of 16 lines of pre-processor, of which 7 are
dedicated to error checking.

This change means that ibv-conduit detects EOP counter equality at the
time that the CQE is processes, and will therefore be capable of
triggering dependent ops from the progress engine (when we get that
far).