ibv: improved EOP counter handling
When an EOP is used for a put_nb, get_nb, or for management of LC for
a Long payload, it is necessary to ensure that the counters for the
initiated and completed operations cannot become equal prior to the
injection of the final ibv-level operation.

This commit moves responsibility from the top-level functions in
gasnet_core.c and gasnet_extended.c, down into the gasnet_rmda_*()
functions in gasnet_core_sndrcv().  As a result the code in the first
two files is simplified, *and* we can now optimize for two cases (the
second of which we could not have previously):
 1. Gets no longer require any extra work to avoid premature balance
 2. Puts that use the inline send also avoid the extra increments

The callers in the Long and put_nb path *still* remain responsible for
the case that no increments were performed at an LC counter.