Initial implementation of GASNETI_FLAG_LC_OPT_IN
This (internal) flag allows using the lc_opt argument to Long
Requests and Replies as a IN paramenter, naming the LC leaf event
of either an eop or iop.

Currently only ibv-conduit includes any asynchronous LC, and thus
the mojority of the changes occur there.  The other EX-enabled
conduits only need to check the flag bit to prevent overwritting
the caller-provided *lc_opt with GASNETEX_INVALID_HANDLE.