aries/gemini: first prereq for IMMEDIATE support
This commit makes the first change needed to support the IMMEDIATE
flag in aries and gemini conduits.

All network operations in aries- and gemini-conduits need a "post
descriptor", allocated using gasnetc_alloc_post_descriptor().  When
there are no post descriptors left, one must spin until one becomes
available.  We do not grow the pool because it lives in the aux seg.

This commit adds a "flags" argument to gasnetc_alloc_post_descriptor()
and modifies its behavior to return NULL if the IMMEDIATE flag is set
and the pool of post descriptors is empty.

Currently all callers are updated to pass 0 for the flags argument.
No callers are yet prepared for a NULL return.