Cray support: use explicit -L/-l for Cray's libs
Prior to this commit we have been dependent on the fact that the cc
and CC wrapper-compilers provided by Cray would "always" link the
required libs (ugni, pmi, xpmem and udreg) implicitly as long as their
corresponding environment modules were loaded.

Well, it turns out that "always" is defined as "when not using dynamic
linking" (which *was* always back when dynamic linking was not
supported by Cray at all).  Today, if one loads the "dynamic-link"
environment module on Titan, or the "cray-memkind" module on Gerty
(among others) then the wrapper-compilers cease to provide the
necessary -L and -l flags for some or all of those libraries because
they become redundant due to shared-library dependencies among Cray-
provided libraries.  However, they are *not* redundant for the link of
GASNet executables, since we don't link any MPI symbols (for instance)
that would trigger the search of shared-lib dependencies.

We recently (March) added support for building on Cray systems
*without* use of the wrapper-compilers, which involved the conditional
use of explict -L and -l options for the Cray-provided libraries that
we use.  So, having confirmed that the implicit and explicit linking
don't interfere, this commit makes the use of explicit -L and -l
options for Cray-provided libraries *UN*conditional.

(cherry picked from commit a79b9df9743801586db1cec9c2a8a83cf0a7059a)