Refine implementation of GASNETI_FLAG_LC_OPT_IN
This commit acts on feedback Dan gave on the initial implementation,
and effectively reverts all of the associated changes made to conduits
other than ibv.  In the case of ibv, the changes relative to the
initial implementation are small.

The major difference is a change to the "contract" with a caller who
passes GASNETI_FLAG_LC_OPT_IN.  Where previously the conduits took
care to never overwrite a caller-provided *lc_opt, it is now permitted
to overwrite it with GASNETEX_INVALID_HANDLE in the case of syncronous
local completion.  This allows for the removal of all the branches the
initial implementation had added to conduits which lack asynchronous
LC support (which means all conduits except ibv).

In the case of ibv-conduit, a branch is still needed to get the
correct behavior for synchronous LC, which differs for a
caller-provided EOP vs one allocated in the current function.