psm: allow runs with GASNET_SUPERNODE_MAXSIZE set
It is *not* safe for psm-conduit to disable psm2's shared-memory comms
just because PSHM is enabled.  If the env var GASNET_SUPERNODE_MAXSIZE
is less than the procs/node, then psm2 *does* need to reach processes
on the same node.

This commit disables the setenv() call that disables psm2 over
shared-memory.  In the future, it may be possible to examine the
nodemap info to determine exactly when shared-memory comms are or are
not needed.