ibv: fix "payload-tearing" bug in IBDISSEM barrier
This commit ensures correct barrier operation on Omni-Path HCAs by
reverting a portion of the the IBDISSEM barrier's specialization
relative to the RDMADISSEM barrier on which it is based.

Prior to this commit, the IBDISSEM barrier made two significant
changes relative to the conduit-independent RDMADISSEM barrier.  One
of these was to replace the redundantly-encoded 128-bit barrier
payload with a simpler 64-bit one.  In addition to halving the
payload size, this enabled use of scalar operations (by value) where
structs were used before (by reference).

However, the use of the 64-bit payload was based on the assumption
that it would always be delivered atomically.  This had not been a
problem until testing of ibv over Omni-Path HCAs showed numerous
barrier mismatches.  These failures were occuring because the delivery
of the 64-bit barrier payload is not atomic.  By reverting to use
of the 128-bit barrier payload of the RDMADISSEM version, the IBDISSEM
barrier no longer makes any unsafe assumptions about the atomicity or
order of delivery of the barrier payload.

The only remaining (meaningfull) difference relative to RDMADISSEM is
the use of an internal Put routine that allows fire-and-forget, thus
eliminating the need to manage (thread-specific) handles.  While this
difference remains, it now represents on the order of only a dozen
lines of code.  So, the "algorithm" really should be factored away
from the "comms", which is the only real conduit-specific logic at
this point. (Dan and I have identified the need to factor this logic
before, but this is not the time for that large a change).

While not directly related to the bug fix, this commit also moves a
call to gasneti_pshm_addr2local() out of the critical path and into
the initialization code.  Note that this path is not even reachable
unless PSHM is ENabled but the hierarchical (pshm+network) barrier
logic has been DISabled (by setting GASNET_PSHM_BARRIER_HIER=0 in the
environment).