psm: provide a "safety net" for bug 3333
This commit adds logic to the gasnet_init() path to detect condtions
which would require use of PSM2's shm device.  If those conditions
are met, then the job will exit by default with a message directing
the user to the documentation for bug 3333, to (soon) appear in
psm-conduit/README.  That documentation will describe an environment
variable setting which disables this entire check, and thus allows use
of PSHM2's shm device.

The point of all of this is to ensure that an end-user will not
encounter the errors reported in bug 3333 unless they "opt in" by
setting the documented env var.

It is also worth noting that the conditions for use of PSM2's shm
device are *not* satisfied by a default, since one must disable
GASNet's own shared-memeory bypass (PSHM) at either configure time
or runtime.  So, the expected common case is that this new logic
adds a very small delay at startup but otherwise does not impact
the end user in any way.