psm: work-around bug 3342
This commit disables the MQ-based protocol for Puts in psm-conduit, in
order to avoid the issue identified in bug 3342.  There is no change
to Gets.

This is intended as a short-term solution to allow timely release of a
conduit which yields correct results.  The longer-timer plan is to
repair the MQ-based protocol instead of disabling it.