Explicitly blacklist -Wno-unused-parameter
'clang -Wall -Wextra -Wno-unused' on clang 3.8.1 was observed to
still generate Wunused-parameter warnings. This is likely due to the
special-case defaulting behavior of this sub-option in the combination
of '-Wall -Wextra'.

Add an explicit disable of this sub-option (when supported) to prevent this problem.

Adjust the configure diagnostic pragma probe to pass exactly the
(preprocessed) code we pass from gasnet_basic.h.

(cherry picked from commit 0d9dd257746717266108b4114b2d56d9866325cf)