Factor conduit-independent AM registration code
This commit factors the conduit-independent portions of AM
registration, in paticular the logic which deals with assignment for
"don't care" entries and detection of duplicates.  The code to
actually "register" remains in the conduit, even though it is a
simple assignment in most cases.

There is significant revision comming to the the registration logic
to accomodate new capabilities in GASNet-EX.  However, this commit
makes NONE of those changes - it only centralizes the code that
will be the home to many of those changes.