Adjust AMX standalone harness tests
Tests 'amxdist' and 'distcheck' in the AMX suites are now off by
default, as they are irrelevant to normal GASNet users and have the
potential to generate irrelevant noise on exotic configurations. They
can optionally be enabled using:

  harness -add_feature=amx_standalone

Adjust test 'tests' to only build when AMX is not the network, to
ensure compile failures on individual tests are better attributed to the
appropriate test when AMX is the network.

Fix an amudprun compile failure on Cygwin that was masked by the above.

(cherry picked from commit 253ea04653b37a2ce07a522df35cc4ba213f0455)