Change dontcare AM handler assignment algorithm
This commit changes the algorithm used to assign AM handler indices to
"dontcare" (index 0 on input) entries to work from the highest available
index to the lowest.  This (when documented) will permit EX clients to
mix subsystems which include fixed and dontcare AM table entries in a
way that one can reason about.