Convert all internal AM tables to EX
This commit removes all internal use of gasnet_handlerentry_t,
replacing it with gasnetex_handlerentry_t.  To retain support for
gasnet_attach() a new gasneti_amregister_legacy() function is added
to perform on-the-fly conversion.

Additionally, this commit factors the instances of the amref handlers,
previously appearing independently in extended-ref/gasnet_extended.c
and in 4 conduits (aries/gemini, ibv, pami, and smp)