Factor constants partitioning the AM handler table
This commit establishes one central definition (in gasnetex.h) of the
constants defining the boundaries between Core, Extended and Client
portions of the AM handler table.

If it should become necessary, a conduit can still provide overrides
in their gasnet_core_fwd.h.

Additionally, gasneti_amregister_legacy() is updated to always use the
range reserved to the client and to perform both the fixed and
dontcare passes.  This eliminates 4 arguments and reduces the required
logic in each conduit's gasnet_attach() implementation).