aries/gemini: avoid an exit race leading to SIGBUS
This commit adds some more synchronization among the processes within
a PSHM supernode to ensure that all reads from the "aux" segment in
the exit coordination code will complete before the lead process in
the supernode can exit (taking the segment with it).

This fixes a real race, which may or may not be responsible for some
of the SIGBUS failures seen in bug 3445.  Note, however, that only
"eventually-collective" exits are ensured the be free of SIGBUS.