aries/gemini: drain sys_exit completions
This commit attempts to ensure that all exit-coordination messages
sent by gasnetc_sys_exit() are "globally complete" before exiting.

Note that a significant portion of the changes are required just
to get a counter associated with the "notify" operation.  The
code to wait on that counter in gasnetc_sys_fini() is only about
half the code.

This commit also removes the prototype for gasnetc_send_control() from
gasnet_gemini.h, rather than updating it, since it has no callers
outside of gasnetc_gemini.c.