Address bug 3346 - x86*/Linux timer calibration
This commit addresses bug 3446 by implementing new TSC calibration
logic for Linux on x86, x86-64 and MIC/k1om which uses the OS provided
wallclock time as a reference.  This new logic is intended to be
highly reliable and thus is being deployed as the default.  However,
environment variables are provided to use the old logic, or to
manually provide the TSC rate; as well as to adjust the allowable
uncertainty in the calibration.

This commit is described as "addresses" rather than "fixes" or
"resolves" because it is still hoped that we may return to use of
/proc/cpuinfo for older CPUs in order to reduce the risk that
overcommit (or noise in VMs or emulated systems) will prevent
calibration.

(cherry picked from commit 171a6b70ed8c9be0241dcd24ad857d2a16af5d28)