ibv: honor RLIMIT_MEMLOCK by default
This commit changes the default behavior of the probe for max pinnable
memory to apply the RLIMIT_MEMLOCK value prior to probing.  Since we
expect that the O/S and/or driver will be enforcing this limit anyway,
we can save time (and confusion) by lowering our expectations
accordingly.  Additionally, this makes setting GASNET_PHYSMEM_NOPROBE
marginally safer.