Implement/factor gasneti_getenv_early()
This commit splits gasneti_getenv() into two parts, one of which is
gasneti_getenv_early() which makes no tracing calls (and so also must
exclude the envdecode hook).

This is intended for use in breaking the cycle between timer and
tracing initialization, and is unlikely to have a more general use.
So this function is currently declared static.