Add debug output for TSC calibration failures
This commit adds more info to errors and warnings regarding a failure
to calibrate the TSC against wallclock.  While this info may have very
little meaning to the end-user, when included in a bug report or email
it can allow the developers to access the degree to which the
calibration was off - and thus determine if adjusting of TOLERANCE env
vars is a sensible recommendation.