tools: build-time control over TSC calibration
This commit adds build-time control over all of the key parameters
which are used to calibrate the x86 TSC against the OS-provided
wallclock time.

The related comments are updated to clarify that they describe the
behavior with the default parameters.

This commit does *NOT* change the default calibration parameters.
However, it introduces a new parameter (TSC_WC_SIGMA) which changes
both how the convergence condition is calculated and reported.
However default the value of the new parameter is 1, resulting in
no change to the default convergence condition.

(cherry picked from commit 6368a1c5732c85ee12bb055d96e90925309fb233)