tools: improved TSC calibration arithmetic
This commit replaces the naive algorithm for online computation of the
mean and variance with one known to be more numerically stable.  In
particular the algorithm used previously is known to suffer from
catistrophic cancellation.

Additionally this commit refines (again) the output for calibration
failures to separate the Sigma from the Tolerance.