Implement lo/hi algorithm for calibrating ticks
This commit implements a new algorithm for calibration of a ticks
source against an OS-provided wallclock, which replaces the previous
"wallclock" algorithm for calibrating the TSC on x86*/Linux.

The new algorithm is based on upper- and lower-bounds which are
collected using sampling methodologies with one-sided errors, and
combined via MIN() and MAX() to yield an estimate that is nearly
oblivous to noise and outliers.

(cherry picked from commit 3fa260cf751f9ad5ef316a83467f5cc16f20a32a)