ofi: Set a limit on the number of AM send buffers
Previously, the conduit was allocating a buffer every time the pool was
accessed while empty. This could lead to high memory usage. This commit
limits the number of allocated buffers and adds polling to wait for a
buffer to become free if a buffer pool is completely consumed.

The number of buffers allocated are configurable by environment variables.