Remove PTHREADS_{INCLUDE,LIB,PATCH} configure opts
Retire the GASNET_CHECK_OVERRIDE_PTHREADS macro, which has outlived its utility.

The PTHREADS_{INCLUDE,LIB} variables are subsumed by CPPFLAGS and
LDFLAGS, respectively, that can be used to select a non-default system
library replacement (and it need not be pthreads-related).

PTHREADS_PATCH has not been used for a long time, and a user who wants
to patch system headers can accomplish the same behavior by patching
them manually and using CPPFLAGS to point GASNet and clients at the
directory where the patched headers are installed.