Implement --disable-munmap (for bug 3480)
This commit is a work-around for
  Bug 3480 - Random memory allocation failures with CLE6 kernels

It has been observed with the compute node kernels on NERSC systems
with CLE6, that there may be a delay between an munmap() call and the
ability to realocate the released page frames.

Both mmap() and malloc() calls have been observed to fail when they
should not.  This can be a serious problem (especially the mmap()
failures) causing "random" failures at startp when using a large
fraction of physical memory.

To commit allows reliable operation in the presence of this behavior
by avoiding mmap() calls prior to creation of segments, thus
eliminating the need to ever call munmap().

This does not actaully make munmap() calls illegal, nor render them
no-ops.  So, conduits which may unmap memory at exit time may
continue to do so.

The only conduits making direct calls to gasneti_{,huge_}munmap()
other than at exit are ibv and mxm when probing for the max
pinnable limit.  When configured using --disable-munmap, these
probes are disabled regardless of the setting of the respective
GASNET_PHYSMEM_{NO,}PROBE environment variable.