Warn users when bug 3480 looks likely
This commit adds logic to detect likely instances of bug 3480, and
issue a warning message to the user.  The logic is based on the
observation that the fixed mmap() calls are always performed right
after an munmap() of equal or larger size.  Thus, in the absence
of external consumers of memory, a failure in a fixed mmap() call
is most likely attributable to bug 3480.