pami: fix a collectives regression
This commit corrects a regression introduced in
  * 52af3a3 pami: remove collectives macros from extended_fwd

Due to header ordering, that commit resulting in code in the
pami-conduit and extended-ref directories disagreeing on whether
or not the PAMI collectives where being used.  This resulted in
NOT using the accelerated collectives, and a SEGV in testteam.

(cherry picked from commit 70228cbfeaa6ac389930a53d2063fcbce53ed364)