bug3480: new work-around part 1-of-2: barrier
This commit adds (subject to --enable-bug3480-workaround) a barrier
between the munmap() of the "protosegment" created by gasnet_Init()
and the mmap() of the final segment requested by gasnet_Attach().
This has been confirmed to eliminate the observed incidence of failed
mappings due to bug 3480 within calls to gasnet_Attach().

This fix is not effective at preventing mmap() failures due to bug
3480 in calls to gasnet_Init() (within mmapLimit() or segmentInit()),
because those are not collective in their use of munmap()/mmap().