bug3480: new work-around part 2-of-2: retries
This commit adds (subject to --enable-bug3480-workaround) bounded
retry of failed mmap() calls which use MMAP_FIXED.  This has been
confirmed to eliminate the observed incidence of failed mappings due
to bug 3480 within calls to gasnet_Init().

This fix is not effective at preventing mmap() failures due to bug
3480 in calls to gasnet_Attach() because application of retry in that
case appears to produce a "live lock" in which the constant retries
are leading to a continuation/extension of the transient leak of
physical page frames.