configure: replace non-portable "test -e"
This commit replaces use of non-portable "test -e" in the pami-conduit
configuration logic with "test -d" (as appropriate to all of the tests
which are seeking either and include or lib directory).  Note that
test does follow symlinks.

Also fixed one instance of "test [DIRNAME]" with no operator after
"test".