fix bug 3498: Sequence-point warnings from --enable-threadinfo-opt PAR mode
Replace the problematic assignment expression with a call to an inline function that
performs the equivalent assignment and adds a sequence point, thereby
avoiding C's undefined behavior and silencing the warning.

In compound gasnet expressions (such as those originally triggering the warning),
the unspecified order of argument evaluation in C means it's unspecified
which expansion of GASNET_GET_THREADINFO() will be evaluated first, but
this non-determinism is not erroneous and is indeed harmless in this
case (as the end result is the same either way).