bug3480: revert most of cb2da1b8 and cead834d
This commit reverts the code changes made in
   cb2da1b8 Implement --disable-munmap (for bug 3480)
and subsequently modified by
   cead834d Rename the configure option to work around bug3480
effectively removing the work-around for bug 3480 while retaining the
supporting configure-related logic.  This change is in preparation for
an entirely new approach to be implemented in the following commits.

(cherry picked from commit fbe1b6b5213011762acdc518ccafa2cc145cce1e)