fix bug 3495: broken pipe message from GASNET_GET_AUTOCONF_VERSION
Use AC_AUTOCONF_VERSION when available (2.62+) to speed the macro.
Remove the problematic and extraneous cat process when it's not.

Remove $AUTOCONF_VERSION output which has no current uses and would
not be reliable for its likely intended purpose of version comparison.
The right way to do this is AC_PREREQ (for a hard requirement) or
m4_version_compare() otherwise (although the latter has its own
portability issues).

$AUTOCONF_VERSION_STR output remains, and its correct usage is
informational for human consumption.