Tweaks to portable timers
bug3508: When the user configures with --enable-force-gettimeofday or
--enable-force-posix-realtime, ensure we disable compilation of
(unreachable) native timer code, in case they are working around a system
that gets compile errors on the native timer code.

Also, fix an ancient typo that was causing the config string to
mis-identify timers_forced_posixrt as timers_posixrt.

Finally, now that configure is unconditionally testing for
clock_gettime, promote that to the default portable timer implementation
when no native implementation is available (eg on a new system), since
it generally outperforms gettimeofday().