Split client/aux segments in EX-supported conduits
This commit updates all of the EX-supported conduits to allocate (and
register when appropriate) the aux segment separately from the client
segment.

The PSHM support for the dual-segment universe is currently less than
ideal, and will benefit (fewer comparisons and branches) from the
eventual use of flags indicating that remote addresses are
IN_AUX_SEGMENT or IN_BOUND_SEGMENT. However, that is not the top
priority yet.

In the cases of PAMI and Aries the RDMA keys for the two segments are
handled relatively gracefully, though of course the mechanism will
need to be generalized and refined for true multi-segment support. As
with PSHM, we'll be able to reduce comparisons and branches when flags
can point us at the proper segment (will still branch for now, until
we get to the point of doing compile time static optimizations).
In the case of UDP and MPI, the expedient course was taken: register
"everything" as the segment for the AMX layer. [FWIW: restoring the
registration of just the client segment provides a means to track down
the places in conduit-independent code that should use the
IN_AUX_SEGMENT flag.]

In the case of IBV the newly added support for distinct registration
keys for aux and client segments is "dirty" and took as long to
implement as all the rest combined. This reflects the fact that the
pre-existing design already deals with the possibility that the client
segment is covered by multiple ibv-level registrations. The addition
of one additional registration for the aux seg has been kludged in
(via negative "rkey_index" or "lkey_index"). However, the
multi-segment support will need a better generalization.